requirements.txt
# The installation of packages is not working consistently right now, this is due conda version conflicts
# please install the packages using conda or pip, if using conda install collecting from conda-forge.
# plotly and dash are not needed for the tool itelf but in order to be able to run the dashboard they are necessary
# platform: win-64
pyyaml~=5.3.1
pandas~=1.1.3
matplotlib~=3.3.2
seaborn~=0.11.0
numpy~=1.19.1
scipy~=1.5.0
statsmodels~=0.12.0
scikit-learn~=0.23.2
powerplantmatching~=0.4.8
cartopy~=0.18.0
geopandas~=0.8.1
requests~=2.25.1
